pyproject.toml
[tool.poetry]
name = "meeting-website"
version = "0.1.0"
description = ""
authors = ["ssyucfa <patin123321@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
Pillow = "^9.1.0"
mixins = "^0.1.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
